move A, Header, Footer out
