+++ src/components/A.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+
+function A(props) {
+  return (
+    <a target="_blank" rel="noopener noreferrer" {...props}>
+      {props.children}
+    </a>
+  );
+}
+
+export default A;
